Add a binary to list recent references
There are a variety of ways to ways to list references, and git
for-each-ref provides several options. However, listing references in
order of recency is not as simple. Add a utility to list references in
terms of recency, both in terms of checkout date (according to the
reflog) and commit and author date.