Makefile: properly detect absence of clippy
When using rustup, such as in the nightly container, if clippy is
missing, the cargo-clippy binary in the PATH will still exist, but it is
just a wrapper for the one in the rustup version directory.

However, when clippy is missing, attempting to invoke that wrapper, even
with "--help", will fail, so we can use this to detect if clippy is
actually present and skip running it if it is not.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>