Add manual pages for each command
It's often useful to have a manual page in addition to help output, so
provide manual pages for each command and build them as part of the CI
process.

Use lower case in the AsciiDoc source to make it easier to read;
Asciidoctor will automatically convert it to upper case when building
the manual page.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>