pktline: avoid returning more bytes than written
The API for std::io::Write is to return the number of bytes that were
passed in from the buffer that were written, not the total number
written to the underlying stream.  If we pass the latter, we end up
exceeding the size of a slice and only partially writing some data, so
let's not count the header as part of the number of bytes written.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>