errors: add functions to find the kind of a error
We have several different kinds of errors and we'll want to choose
different behaviors based on the kind.  Let's provide kind and io_kind
methods to find out our ErrorKind and the std::io::ErrorKind for our
errors.  Let's use this function down below when converting into a
std::io::Error as well.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>