Add a basic implementation of the Git LFS SSH transfer server
It's helpful to have a basic implementation of the Git LFS transfer
server so we can verify that the design works as expected and is
reasonably performant.  Since the protocol is reasonably simple, let's
implement it in a basic server implementation.

The server does not support locking, but it does support uploads and
downloads to a remote Git repository.  It should theoretically work on
Windows, but this has not been tested.

Disable the new clippy lint that came in with 1.47.0, since that does
not work on our oldest supported version, 1.41.1.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>