git-lfs-transfer: ignore unnecessary-wraps clippy lint
Clippy wants us to avoid using a Result when we can only have one
possibility (in this case, success).  However, this ignores the fact
that we want to structure our code in a parallel way such that all
similar methods return similar values.  As a result, this warning is
unhelpful, since it moves us away from that direction, so let's disable
it.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>