git-test-merge: avoid needless question mark operator
In both these cases, we have a function which returns a Result, but we
use the question mark operator inside an Ok.  That's silly, since
they're equivalent but what we've written is more verbose and harder to
reason about, and Clippy rightfully warns about it.  Let's fix this by
simplifying our code so it's easier to understand.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>