Makefile: add targets to build Markdown
When we upload a crate to crates.io, it wants the README to be in
Markdown, but we prefer AsciiDoc for documentation.  To help sort this
out, let's add a target to build Markdown files from AsciiDoc files and
let's update the .gitignore and clean targets to ignore and clean up
these files.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>