git-lfs-transfer: fix return type on non-Unix
The Unix version of the set_permissions function returns a Result<u32,
Error>.  Because we save the u32 result later on in generic code, our
non-Unix function must return the same thing.  Note that on non-Unix
platforms the value is not used, so while we must provide something, it
doesn't matter much what.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>