git-lfs-transfer: add locking support
Add support for basic locking and unlocking.  Store our locks as a flat
file with metadata inside it with a name that's a hash of the path name
and a version number for easy lookup.

Unlike many forges, we don't restrict access to lock and unlock files;
if the user has permissions to operate on the remote repository, then
we'll perform the operation.

Note that due to lack of support in the standard library, we currently
have very poor support for owner information.  On Windows, no user
information is provided, and on Unix, we emit only information by user
ID.  This will be fixed in a future release, but the current
implementation provides the minimum viable feature.