pktline: flush stream after flush packet
In pkt-line-based protocols, a flush packet means the message is
complete.  Oftentimes, that means that we are waiting for the other side
to take our message into account and act accordingly.

It is not very helpful in getting the other side to do this if they do
not receive the flush packet, as we have failed to tell them it's their
turn.  Notably, in git-lfs-transfer, we fail to interoperate with the
other side since we fail to flush the stream.

So, to help both sides communicate effectively, let's flush the stream
after sending a flush packet.  Since flush packets are relatively
uncommon, there should be no interesting performance consequences from
this.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>