git-lfs-transfer: use size parameter for verify-object
The current implementation of verify-object uses the original spec,
which specified the size as part of the command.  However, that behavior
was inconsistent with how get-object and put-object behave, and the spec
is being updated to make verify-object consistent.

Let's update the verify-object implementation to use the same parsing as
the put-object implementation and additionally gracefully handle
unexpected arguments.  Simplify it slightly to use the slice's
starts_with method instead of hard-coding the implementation.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>