scutiger-bin: fix tests with non-default branch name
When a user has init.defaultBranch set, that results in our test
repository having HEAD set to that instead, breaking the tests.  Let's
handle this by explicitly initializing the repository to use a
non-default name (dev) and using it everywhere in the testsuite.

This ensures that neither the old default of master nor the future
default of main are the default and means our CI will test this properly
in all cases.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>