pktline: fix reader with large messages
If the message size is large than a single packet, in some cases our
reader doesn't work.  That's because we fail to handle the leftover data
after we've filled the buffer to capacity.

Let's fix this by making sure we update the length field in our reader
to reflect the data we've copied and adding an offset field.  We could
avoid the need for the offset field by performing an additional copy,
but since our buffer is likely to be a power of two and the amount of
data often is not (since we must strip off the pkt-line header), we'd
perform an additional copy in almost every case, which would be much
less efficient than using the offset field.

To make sure we're copying the data we expect, let's add a test function
that generates a changing pattern and use it whenever we're dealing with
our buffer.  We specifically choose some test chunk sizes that are prime
(17 and 32771) to make sure we're exercising all cases.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>