scutiger-lfs: move locking code to local backend
Move the locking code to the local backend and add traits for locking to
the backend.  Since the override of user_for_file in test mode is no
longer possible due to the library being a separate crate, replace the
strings used in the locking tests with the correct value.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>