Remove atty dependency
We want to support older versions of Rust, but atty has a potentially
unaligned read, and the only versions of clap that support our desired
versions of Rust use atty for colour.  We'll never actually hit this in
practice, since it requires a custom allocator, but it causes some
security warnings.

As a result, let's disable the default features to get rid of the atty
dependency.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>