Use Git-based repository for git-lfs-transfer
Now that we have a Git-based repository, use it for `git-lfs-transfer`.
This adds SHA-256 repository support and also allows us to process
repositories that are not owned by the current user.

Note that because we call a bare program name, it's not safe to use this
software on Windows because of a Rust vulnerability.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>