git-lfs-transfer: set executable bit on directories
If `core.sharedrepository` is set, we set the umask such that the
executable bit is always cleared.  Creating directories this way is not
especially useful, because it means that nobody can search inside of
them.

Let's fix that by automatically adjusting the permissions to set the
executable bit if the corresponding read bit is set, which means that
directories will either have both or neither.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>